fix: bug fixes, regression tests, and README improvements
Bugs fixed:
- evaluator: add NE to isBooleanOP (string != was returning UNDEFINED_RESULT)
- evaluator: guard MOD-by-zero for both Int and Float
- collectjs: MergeBy propagates nv.err instead of nil err variable
- collectjs_types: NULL_RESULT.datatype = Null (was incorrectly Undefined)
- parse: Expr() returns nil on empty stack instead of panicking
- parse: ExitXpath_name guards empty string before str[0] access
- context: absFuncHandle uses float64 sign check for FloatNode

Test fixes:
- Fix malformed JSON and syntax errors in evaluator_test.go
- Fix flawed test comparison condition (type check was inverted)
- Fix pre-existing wrong expected values in collectjs_test.go
- Rename malformed Example_ functions for Go 1.24 compatibility

New tests (evaluator_regression_test.go):
- TestModByZero, TestNEOperatorWithNonNumericStrings, TestNEOperatorWithNumbers
- TestAbsFunction, TestNotOperator, TestCaseWhenExpr, TestNullResultType
- TestEmptyStackExpr, TestArithmeticEdgeCases, TestBooleanComparisons
- TestUndefinedInExpressions, TestSQLSelectWithAlias

Documentation:
- Rewrite README.md: correct title, features, quick start, syntax reference,
  API docs, and architecture diagram

+++ context.go
@@ -77,7 +77,7 @@ var absFuncHandle = func(args ...Node) Node {
 			}
 			return arg
 		case FloatNode:
-			if int64(arg) < 0 {
+			if float64(arg) < 0 {
 				return -1 * arg
 			}
 			return arg
